+++ app/src/main/java/com/andrerinas/openheadunit/aap/AapSsl.kt
@@ -1,11 +1,11 @@
 package com.andrerinas.openheadunit.aap
 
-import com.andrerinas.openheadunit.connection.AccessoryConnection
+import com.andrerinas.openheadunit.connection.projection.ProjectionConnection
 
 interface AapSsl {
     fun decrypt(start: Int, length: Int, buffer: ByteArray): ByteArrayWithLimit?
     fun encrypt(offset: Int, length: Int, buffer: ByteArray): ByteArrayWithLimit?
     fun postHandshakeReset()
-    fun performHandshake(connection: AccessoryConnection): Boolean
+    fun performHandshake(connection: ProjectionConnection): Boolean
     fun release()
 }
