--- app/src/main/java/com/andrerinas/openheadunit/connection/AccessoryConnection.kt
@@ -1,10 +1,0 @@
-package com.andrerinas.openheadunit.connection
-
-interface AccessoryConnection {
-    val isSingleMessage: Boolean
-    fun sendBlocking(buf: ByteArray, length: Int, timeout: Int): Int
-    fun recvBlocking(buf: ByteArray, length: Int, timeout: Int, readFully: Boolean): Int
-    val isConnected: Boolean
-    suspend fun connect(): Boolean
-    fun disconnect()
-}
